Careers/requirements.txt
asgiref==3.7.2
boto3==1.26.117
botocore==1.29.117
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.2.5
django-active-link==0.1.8
django-cors-headers==3.13.0
django-phonenumber-field==7.0.1
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
mysql-connector-python==8.0.33
phonenumbers==8.13.4
Pillow==10.1.0
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2023.3
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
